+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `omnipytent.ext.extra.screen` module for interaction with [GNU Screen](https://www.gnu.org/software/screen/)
 - `omnipytent.ext.extra.testing` subpackage for test choosing helpers.
 - `omnipytent.ext.extra.repls` subpackage for interacting with REPLs. Currently has one submodule for `ipython`
+- `omnipytent.ext.extra.input` subpackage for extending input utilities.